deprecate(rhino): removes Rhino BIM functionality (#2602)
deprecates rhino bim from connector, converter, and toolbar
