+++ Objects/Converters/ConverterRhinoGh/ConverterRhinoGhShared/ConverterRhinoGh.BuiltElements.cs
@@ -17,6 +17,7 @@ namespace Objects.Converter.RhinoGh;
 
 public partial class ConverterRhinoGh
 {
+  // views
   public View3D ViewToSpeckle(ViewInfo view)
   {
     // get orientation vectors
@@ -141,271 +142,7 @@ public partial class ConverterRhinoGh
     return viewport;
   }
 
-  public Column CurveToSpeckleColumn(RH.Curve curve)
-  {
-    return new Column(CurveToSpeckle(curve)) { units = ModelUnits };
-  }
-
-  public Beam CurveToSpeckleBeam(RH.Curve curve)
-  {
-    return new Beam(CurveToSpeckle(curve)) { units = ModelUnits };
-  }
-
-  // args of format [width, height, diameter]
-  public Duct CurveToSpeckleDuct(RH.Curve curve, string[] args, out List<string> notes)
-  {
-    Duct duct = null;
-    notes = new List<string>();
-
-    if (args.Length < 3)
-    {
-      notes.Add("Number of schema arguments less than 3");
-      return duct;
-    }
-    if (
-      double.TryParse(args[0], out double height)
-      && double.TryParse(args[1], out double width)
-      && double.TryParse(args[2], out double diameter)
-    )
-      duct = new Duct(CurveToSpeckle(curve), width, height, diameter)
-      {
-        units = ModelUnits,
-        length = curve.GetLength()
-      };
-    else
-      notes.Add("Could not parse schema arguments into doubles");
-    return duct;
-  }
-
-  // args of format [diameter]
-  public Pipe CurveToSpecklePipe(RH.Curve curve, string[] args, out List<string> notes)
-  {
-    Pipe pipe = null;
-    notes = new List<string>();
-
-    if (args.Length < 1)
-    {
-      notes.Add("Number of schema arguments less than 1");
-      return pipe;
-    }
-    if (double.TryParse(args[0], out double diameter))
-      pipe = new Pipe(CurveToSpeckle(curve), curve.GetLength(), diameter) { units = ModelUnits };
-    else
-      notes.Add("Could not parse schema arguments into doubles");
-    return pipe;
-  }
-
-  public Opening CurveToSpeckleOpening(RH.Curve curve)
-  {
-    return new Opening((ICurve)ConvertToSpeckle(curve)) { units = ModelUnits };
-  }
-
-  public Floor CurveToSpeckleFloor(RH.Curve curve)
-  {
-    return new Floor((ICurve)ConvertToSpeckle(curve)) { units = ModelUnits };
-  }
-
-  public Wall BrepToSpeckleWall(RH.Brep brep, out List<string> notes)
-  {
-    Wall wall = null;
-    notes = new List<string>();
-
-    RH.BoundingBox brepBox = brep.GetBoundingBox(false);
-    double height = brepBox.Max.Z - brepBox.Min.Z; // extract height
-    var bottomCurves = GetSurfaceBrepEdges(brep, getBottom: true); // extract baseline
-    if (bottomCurves == null)
-    {
-      notes.Add("Could not extract wall bottom curves");
-      return wall;
-    }
-    var intCurves = GetSurfaceBrepEdges(brep, getInterior: true); // extract openings
-    List<Base> openings = new();
-    if (intCurves != null)
-      foreach (ICurve crv in intCurves)
-        openings.Add(new Opening(crv));
-    if (height > 0)
-      wall = new Wall(height, bottomCurves[0], openings) { units = ModelUnits };
-    else
-      notes.Add("Wall height is 0");
-    return wall;
-  }
-
-  public Floor BrepToSpeckleFloor(RH.Brep brep, out List<string> notes)
-  {
-    Floor floor = null;
-    notes = new List<string>();
-
-    var extCurves = GetSurfaceBrepEdges(brep); // extract outline
-    if (extCurves == null)
-    {
-      notes.Add("Could not extract floor outline curves");
-      return floor;
-    }
-    var intCurves = GetSurfaceBrepEdges(brep, getInterior: true); // extract voids
-    floor = new Floor(extCurves[0], intCurves) { units = ModelUnits };
-    return floor;
-  }
-
-  public Roof BrepToSpeckleRoof(RH.Brep brep, out List<string> notes)
-  {
-    Roof roof = null;
-    notes = new List<string>();
-
-    var extCurves = GetSurfaceBrepEdges(brep); // extract outline
-    if (extCurves == null)
-    {
-      notes.Add("Could not extract roof outline curves");
-      return roof;
-    }
-    var intCurves = GetSurfaceBrepEdges(brep, getInterior: true); // extract voids
-    roof = new Roof(extCurves[0], intCurves) { units = ModelUnits };
-    return roof;
-  }
-
-  public RV.RevitFaceWall BrepToFaceWall(RH.Brep brep, string[] args)
-  {
-    if (brep.Faces.Count > 1)
-      return null;
-
-    string family = "Default";
-    string type = "Default";
-    try
-    {
-      family = args[0];
-      type = args[1];
-    }
-    catch { }
-    return new RV.RevitFaceWall(family, type, BrepToSpeckle(brep), null) { units = ModelUnits };
-  }
-
-  public Topography BrepToTopography(RH.Brep brep)
-  {
-    brep.Repair(Doc.ModelAbsoluteTolerance);
-    var displayMesh = GetBrepDisplayMesh(brep);
-    return new Topography(MeshToSpeckle(displayMesh)) { units = ModelUnits };
-  }
-
-  public Topography MeshToTopography(RH.Mesh mesh)
-  {
-    return new Topography(MeshToSpeckle(mesh)) { units = ModelUnits };
-  }
-
-#if RHINO7
-  public Topography MeshToTopography(RH.SubD subD)
-    {
-      var speckleMesh = MeshToSpeckle(subD);
-      return speckleMesh != null ? new Topography(speckleMesh) { units = ModelUnits } : null;
-    }
-#endif
-
-  public RV.AdaptiveComponent InstanceToAdaptiveComponent(InstanceObject instance, string[] args)
-  {
-    if (args.Length == 0)
-      return null;
-
-    string family = "Default";
-    string type = "Default";
-    try
-    {
-      family = args[0];
-      type = args[1];
-    }
-    catch { }
-
-    var points = instance.GetSubObjects().Select(o => PointToSpeckle(((RH.Point)o.Geometry).Location)).ToList();
-
-    var adaptiveComponent = new RV.AdaptiveComponent(type, family, points);
-    adaptiveComponent.units = ModelUnits;
-
-    return adaptiveComponent;
-  }
-
-  public RV.DirectShape BrepToDirectShape(RH.Brep brep, string[] args)
-  {
-    if (args == null || args.Length == 0)
-      return null;
-    if (!Enum.TryParse($"{args[0]}s", out RV.RevitCategory category))
-      return null;
-    string name = "DirectShape";
-    try
-    {
-      name = args[1];
-    }
-    catch { }
-    return new RV.DirectShape(name, category, new List<Base> { ConvertToSpeckle(brep) }) { units = ModelUnits };
-  }
-
-  public RV.DirectShape ExtrusionToDirectShape(RH.Extrusion extrusion, string[] args)
-  {
-    if (args == null || args.Length == 0)
-      return null;
-    if (!Enum.TryParse($"{args[0]}s", out RV.RevitCategory category))
-      return null;
-    string name = "DirectShape";
-    try
-    {
-      name = args[1];
-    }
-    catch { }
-    return new RV.DirectShape(name, category, new List<Base> { ConvertToSpeckle(extrusion) }) { units = ModelUnits };
-  }
-
-  public RV.DirectShape MeshToDirectShape(RH.Mesh mesh, string[] args)
-  {
-    if (args == null || args.Length == 0)
-      return null;
-    if (!Enum.TryParse($"{args[0]}s", out RV.RevitCategory category))
-      return null;
-    string name = "DirectShape";
-    try
-    {
-      name = args[1];
-    }
-    catch { }
-    return new RV.DirectShape(name, category, new List<Base> { ConvertToSpeckle(mesh) }) { units = ModelUnits };
-  }
-
-  // edge curve convenience method
-  private List<ICurve> GetSurfaceBrepEdges(
-    RH.Brep brep,
-    bool getExterior = true,
-    bool getInterior = false,
-    bool getBottom = false
-  )
-  {
-    double tol = Doc.ModelAbsoluteTolerance * 1;
-
-    RH.Curve[] brpCurves = null;
-    if (getInterior)
-      brpCurves = brep.DuplicateNakedEdgeCurves(false, true);
-    else
-      brpCurves = brep.DuplicateNakedEdgeCurves(true, false);
-    if (getBottom)
-    {
-      var bottomCrv = brpCurves
-        .Where(o => o.IsLinear())
-        ?.Where(
-          o =>
-            new RH.Vector3d(
-              o.PointAtEnd.X - o.PointAtStart.X,
-              o.PointAtEnd.Y - o.PointAtStart.Y,
-              o.PointAtEnd.Z - o.PointAtStart.Z
-            ).IsPerpendicularTo(RH.Vector3d.ZAxis)
-        )
-        ?.Aggregate((curMin, o) => curMin == null || o.PointAtStart.Z < curMin.PointAtStart.Z ? o : curMin);
-      if (bottomCrv != null)
-        brpCurves = new[] { bottomCrv };
-    }
-
-    List<ICurve> outCurves = null;
-    if (brpCurves != null && brpCurves.Count() > 0)
-      outCurves =
-        brpCurves.Count() == 1
-          ? new List<ICurve> { (ICurve)ConvertToSpeckle(brpCurves[0]) }
-          : RH.Curve.JoinCurves(brpCurves, tol).Select(o => (ICurve)ConvertToSpeckle(o)).ToList();
-    return outCurves;
-  }
-
+  // direct shape
   public List<object> DirectShapeToNative(RV.DirectShape directShape, out List<string> log)
   {
     log = new List<string>();
